Лекция3-Задание1_вер2
